Additional fixes on calories and questionary

+++ lib/domain/repo/questions/questions_repo.dart
@@ -3,6 +3,6 @@ import 'package:calora/domain/model/questions/questions_request.dart';
 
 abstract class QuestionsRepo {
   Future<void> sendAnswers(QuestionsRequest answers);
-  Future<List<int>> getDailyGoals();
+  Future<List<double>> getDailyGoals();
   Future<void> sendTargetWeight(NormsRequest weight);
 }
